# Improve highlight performance