--- a/lib/Extension/LanguageServerReferenceFinder/Model/Highlight.php
+++ b/lib/Extension/LanguageServerReferenceFinder/Model/Highlight.php
@@ -4,8 +4,10 @@ namespace Phpactor\Extension\LanguageServerReferenceFinder\Model;
 
 use Phpactor\LanguageServerProtocol\DocumentHighlightKind;
 
-class Highlight {
+class Highlight
+{
     public int $start;
+
     public int $end;
 
     /**