--- a/lib/Extension/LanguageServerReferenceFinder/Model/Highlighter.php
+++ b/lib/Extension/LanguageServerReferenceFinder/Model/Highlighter.php
@@ -43,6 +43,7 @@ class Highlighter
             $offsets[] = $highlight->end;
             $highlights[] = $highlight;
         }
+
         $lineCols = LineCols::fromByteOffsetInts($source, $offsets);
         $lspHighlights = [];
 
@@ -51,8 +52,8 @@ class Highlighter
             $endPos = $lineCols->get($highlight->end);
             $lspHighlights[] = new DocumentHighlight(
                 new Range(
-                    new Position($startPos->line(), $startPos->col()),
-                    new Position($endPos->line(), $endPos->col()),
+                    new Position($startPos->line() - 1, $startPos->col() - 1),
+                    new Position($endPos->line() - 1, $endPos->col() - 1),
                 ),
                 $highlight->kind
             );