… lib/Extension/LanguageServerReferenceFinder/Tests/Bench/HighlightBench.php
… lib/TextDocument/LineCols.php