In large files (e.g. the tolerantp PHP parser) highlighting requests can take up to 500ms. This was due to converting hundreds of byte offsets to line / char positions which involved iterating over all lines in the text document for every single offset.

This PR makes introduces a way to batch convert byte offsets to line / char positions, it makes highlighting 10x faster:

```
1: HighlightBench 

By Benchmark

446.6ms   │   █ 
390.8ms   │   █ 
335.0ms   │   █ 
279.1ms   │   █ 
223.3ms   │   █ 
167.5ms   │   █ 
111.7ms   │   █ 
55.8ms    │ ▆ █ 
          └─────
            1 2 

[█ HighlightBench] 

1: <current> 2: master    

Table
=====

<current>
+----------------+--------------------+----------+----------+----------+----------+--------+---------+
| benchmark      | subject            | memory   | min      | max      | mode     | rstdev | stdev   |
+----------------+--------------------+----------+----------+----------+----------+--------+---------+
| HighlightBench | benchHighlights () | 20.570mb | 40.338ms | 44.741ms | 41.562ms | ±3.25% | 1.373ms |
+----------------+--------------------+----------+----------+----------+----------+--------+---------+

master
+----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
| benchmark      | subject            | memory   | min       | max       | mode      | rstdev | stdev   |
+----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
| HighlightBench | benchHighlights () | 20.576mb | 445.183ms | 456.593ms | 446.634ms | ±0.90% | 4.069ms |
+----------------+--------------------+----------+-----------+-----------+-----------+--------+---------+
```